+++ src/d/d_camera.cpp
@@ -11206,7 +11206,7 @@ static leafdraw_method_class method = {
 };
 
 camera_process_profile_definition g_profile_CAMERA = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e),
     11,
     fpcPi_CURRENT_e,
     PROC_CAMERA,
@@ -11227,7 +11227,7 @@ camera_process_profile_definition g_profile_CAMERA = {
 };
 
 camera_process_profile_definition g_profile_CAMERA2 = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e),
     11,
     fpcPi_CURRENT_e,
     PROC_CAMERA2,
