+++ src/d/d_envse.cpp
@@ -237,7 +237,7 @@ static leafdraw_method_class l_dEnvSe_Method = {
 };
 
 kankyo_process_profile_definition g_profile_ENVSE = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e), // TPPC
     2,
     fpcPi_CURRENT_e,
     PROC_ENVSE,
