+++ src/d/d_gameover.cpp
@@ -498,7 +498,7 @@ static leafdraw_method_class l_dGameover_Method = {
 };
 
 msg_process_profile_definition g_profile_GAMEOVER = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e), // TPPC
     0x0C,
     fpcPi_CURRENT_e,
     PROC_GAMEOVER,
