+++ src/d/d_k_wmark.cpp
@@ -169,7 +169,7 @@ static leafdraw_method_class l_dkWmark_Method = {
 };
 
 kankyo_process_profile_definition g_profile_WMARK = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e),
     9,
     fpcPi_CURRENT_e,
     PROC_WMARK,
