+++ src/d/d_k_wpillar.cpp
@@ -133,7 +133,7 @@ static leafdraw_method_class l_dkWpillar_Method = {
 };
 
 kankyo_process_profile_definition g_profile_WPILLAR = {
-    fpcLy_CURRENT_e,
+    static_cast<unsigned int>(fpcLy_CURRENT_e),
     9,
     fpcPi_CURRENT_e,
     PROC_WPILLAR,
